+++ my_list.rb
@@ -1,0 +1,20 @@
+require './my_enumerable'
+
+class MyList
+  def initialize(list)
+    @list = list
+  end
+
+  include MyEnumerable
+
+  def each(&block)
+    @list.each(&block)
+  end
+end
+
+list = MyList.new([1, 2, 3, 4])
+puts(list.all? { |e| e < 5 })
+puts(list.all? { |e| e > 5 })
+puts(list.any? { |e| e == 2 })
+puts(list.any? { |e| e == 5 })
+print "#{list.filter(&:even?)} \n"
